pyproject.toml
[project]
name = "live-shop-agent"
version = "1.0.0"
description = "AI-powered live commerce agent using Gemini Live API and Google ADK for real-time product detection, inventory queries, and one-click purchases"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk",
    "temporalio>=1.5.0",
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.16.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
    "termcolor>=2.0.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]